analyzeStorage: move fn  in database.lua
