+++ autoSpread.lua
@@ -4,7 +4,6 @@ local gps = require('gps')
 local scanner = require('scanner')
 local config = require('config')
 local events = require('events')
-local storage = require('storage')
 local breedRound = 0
 local emptySlot
 local targetCrop
@@ -143,7 +142,7 @@ local function main()
 
     -- First Run
     spreadOnce(true)
-    storage.analyzeStorage(true)
+    database.analyzeStorage(true)
     action.restockAll()
 
     -- Loop
