+++ uninstall.lua
@@ -10,7 +10,6 @@ local scripts = {
     'autoStat.lua',
     'autoTier.lua',
     'autoSpread.lua',
-    'storage.lua',
     'uninstall.lua'
 }
